[refactor] 어플리케이션 서비스의 인터페이스를 모두 제거합니다.
