--- src/main/java/org/sopt/pawkey/backendapi/domain/Post/application/service/PostServiceImpl.java
@@ -1,72 +1,0 @@
-package org.sopt.pawkey.backendapi.domain.post.application.service;
-
-import java.util.List;
-
-import org.sopt.pawkey.backendapi.domain.image.domain.model.ImageType;
-import org.sopt.pawkey.backendapi.domain.image.infra.persistence.entity.ImageEntity;
-import org.sopt.pawkey.backendapi.domain.pet.infra.persistence.entity.PetEntity;
-import org.sopt.pawkey.backendapi.domain.post.application.dto.command.PostRegisterCommand;
-import org.sopt.pawkey.backendapi.domain.post.domain.repository.PostRepository;
-import org.sopt.pawkey.backendapi.domain.post.exception.PostBusinessException;
-import org.sopt.pawkey.backendapi.domain.post.exception.PostErrorCode;
-import org.sopt.pawkey.backendapi.domain.post.infra.persistence.entity.PostEntity;
-import org.sopt.pawkey.backendapi.domain.post.infra.persistence.entity.PostImageEntity;
-import org.sopt.pawkey.backendapi.domain.routes.infra.persistence.entity.RouteEntity;
-import org.sopt.pawkey.backendapi.domain.user.infra.persistence.entity.UserEntity;
-import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class PostServiceImpl implements PostService {
-
-	private final PostRepository postRepository;
-
-	@Override
-	public PostEntity findById(Long postId) {
-		return postRepository.findById(postId)
-			.orElseThrow(() -> new PostBusinessException(PostErrorCode.POST_NOT_FOUND));
-	}
-
-	@Override
-	public PostEntity savePost(UserEntity writer,
-		PostRegisterCommand command,
-		RouteEntity route,
-		List<ImageEntity> images) {
-
-		if (!route.getUser().equals(writer)) {
-			throw new PostBusinessException(PostErrorCode.POST_WRITE_FORBIDDEN);
-		}
-
-		PetEntity pet = writer.getPetOrThrow();
-		PostEntity post = PostEntity.builder()
-			.user(writer)
-			.route(route)
-			.title(command.title())
-			.description(command.description())
-			.isPublic(command.isPublic())
-			.pet(pet)
-			.build();
-
-		for (ImageEntity image : images) {
-			PostImageEntity postImage = PostImageEntity.builder()
-				.imageType(ImageType.WALK_POST)
-				.image(image)
-				.post(post)
-				.build();
-
-			post.getPostImageEntityList().add(postImage);
-		}
-		pet.incrementWalkCount();
-		postRepository.save(post);
-		return post;
-	}
-
-	@Override
-	public boolean existsByRouteId(Long routeId) {
-		return postRepository.existsByRouteId(routeId);
-	}
-
-}
-
