--- src/main/java/org/sopt/pawkey/backendapi/domain/category/application/service/CategoryService.java
@@ -1,4 +1,0 @@
-package org.sopt.pawkey.backendapi.domain.category.application.service;
-
-public interface CategoryService {
-}
