--- src/main/java/org/sopt/pawkey/backendapi/domain/category/application/service/CategoryServiceImpl.java
@@ -1,4 +1,0 @@
-package org.sopt.pawkey.backendapi.domain.category.application.service;
-
-public class CategoryServiceImpl {
-}
