+++ src/main/java/org/sopt/pawkey/backendapi/domain/category/application/service/SelectQueryService.java
@@ -3,7 +3,21 @@ package org.sopt.pawkey.backendapi.domain.category.application.service;
 import java.util.List;
 
 import org.sopt.pawkey.backendapi.domain.category.application.dto.response.SelectResult;
+import org.sopt.pawkey.backendapi.domain.category.domain.repository.SelectRepository;
+import org.springframework.stereotype.Service;
 
-public interface SelectQueryService {
-	List<SelectResult> getAllSelects();
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class SelectQueryService {
+	private final SelectRepository selectRepository;
+
+	public List<SelectResult> getAllSelects() {
+		return selectRepository.findAllSelectWithOptionsOrderedById()
+			.stream()
+			.map(SelectResult::fromEntity)
+			.toList();
+	}
 }
+
