+++ src/main/java/org/sopt/pawkey/backendapi/domain/region/application/facade/query/GetRegionFacade.java
@@ -1,7 +1,7 @@
 package org.sopt.pawkey.backendapi.domain.region.application.facade.query;
 
 import org.sopt.pawkey.backendapi.domain.region.application.dto.result.GetRegionResult;
-import org.sopt.pawkey.backendapi.domain.region.application.service.query.RegionQueryService;
+import org.sopt.pawkey.backendapi.domain.region.application.service.RegionQueryService;
 import org.sopt.pawkey.backendapi.domain.region.infra.persistence.entity.RegionEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
