+++ src/main/java/org/sopt/pawkey/backendapi/domain/region/application/facade/query/GetRegionListFacade.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.sopt.pawkey.backendapi.domain.region.application.dto.command.GetRegionListCommand;
 import org.sopt.pawkey.backendapi.domain.region.application.dto.result.GetRegionListResult;
-import org.sopt.pawkey.backendapi.domain.region.application.service.query.RegionQueryService;
+import org.sopt.pawkey.backendapi.domain.region.application.service.RegionQueryService;
 import org.sopt.pawkey.backendapi.domain.region.infra.persistence.entity.RegionEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
