+++ src/main/java/org/sopt/pawkey/backendapi/domain/user/application/service/UserService.java
@@ -1,14 +1,52 @@
 package org.sopt.pawkey.backendapi.domain.user.application.service;
 
+import org.sopt.pawkey.backendapi.domain.region.domain.RegionRepository;
 import org.sopt.pawkey.backendapi.domain.region.infra.persistence.entity.RegionEntity;
 import org.sopt.pawkey.backendapi.domain.user.application.dto.request.CreateUserCommand;
+import org.sopt.pawkey.backendapi.domain.user.domain.repository.UserQueryRepository;
+import org.sopt.pawkey.backendapi.domain.user.domain.repository.UserRepository;
+import org.sopt.pawkey.backendapi.domain.user.exception.UserBusinessException;
+import org.sopt.pawkey.backendapi.domain.user.exception.UserErrorCode;
 import org.sopt.pawkey.backendapi.domain.user.infra.persistence.entity.UserEntity;
+import org.springframework.stereotype.Service;
 
-public interface UserService {
-	UserEntity findById(Long userId);
+import lombok.RequiredArgsConstructor;
 
-	UserEntity saveUser(CreateUserCommand command, RegionEntity region);
+@Service
+@RequiredArgsConstructor
+public class UserService {
 
-	void updateUserRegion(UserEntity user, RegionEntity region);
+	private final UserQueryRepository userQueryRepository;
+	private final UserRepository userRepository;
+	private final RegionRepository regionRepository;
+	
+	public UserEntity findById(final Long id) {
+		return userRepository.findById(id)
+			.orElseThrow(() -> new UserBusinessException(UserErrorCode.USER_NOT_FOUND));
+	}
+
+	public UserEntity saveUser(CreateUserCommand command, RegionEntity region) {
+
+		if (userRepository.existsByLoginId(command.loginId())) {
+			throw new UserBusinessException(UserErrorCode.USER_DUPLICATE_LOGIN_ID);
+		}
+
+		UserEntity user = UserEntity.builder()
+			.loginId(command.loginId())
+			.password(command.password())
+			.name(command.name())
+			.gender(command.gender())
+			.age(command.age())
+			.region(region)
+			.build();
+
+		return userRepository.save(user);
+	}
+
+	public void updateUserRegion(UserEntity user, RegionEntity region) {
+		user.updateRegion(region);
+
+	}
 }
+
